+++ src/logic/facade.h
@@ -30,19 +30,19 @@ class Facade {
   static std::shared_ptr<Facade> GetInstance();
 
   // Функции для работы с принтерами
-  bool addPrinter(const QString &name, double power, int age, double cost);
+  bool addPrinter(const QString &name, int power, int age, double cost);
   QList<QString> getPrinterList();
   bool deletePrinterByName(const QString &name);
   void updatePrinterByName(const QString &oldName, const QString &newName,
-                           double power, int weight, double cost);
+                           int power, int weight, double cost);
   QHash<QString, QVariant> getPrinterByName(const QString &name);
 
   // Функции для работы с катушками
-  bool addPlastic(const QString &name, double weight, double cost);
+  bool addPlastic(const QString &name, int weight, double cost);
   QList<QString> getPlasticList();
   bool deletePlasticByName(const QString &name);
   void updatePlasticByName(const QString &oldName, const QString &newName,
-                           double weight, double cost);
+                           int weight, double cost);
   QHash<QString, QVariant> getPlasticByName(const QString &name);
 
   // Функции для работы с пресетами
